Move MapAdapter to top level
